databases/mariaDB/app/requirements.txt
Django
django-jet-reboot
djangorestframework
mysqlclient
six
pycryptodome
requests